requirements.txt
requests
python-dotenv
flask
pytest
dataclasses
RPi.GPIO
